Select Current Family Member.
After login to the family account people should be able to pick who they are.

Used the context to pass person's names to the html which lists them as buttons. Then Post
sets the current person (model.ChooseFacilitator) to that person

The only issue I had was trying to do this with a class bases view as most documentation was
done in function based view