Add links between the user home page and the calendar
Added links between the user Home page and the Calendar page. Also
changed the initial page loaded to the user login page instead of the
calendar page; Added another icon to the sidebar to direct to the
home page and linked that same icon it to the user home page.