Change the Home page to a login prompt instead of a link to the login prompt.
There was a link on the homepage to the login page. There wasn't really
a purpose for that page so I got rid of it and directed the homepage
straight to the login; The original 'Home' url address was changed from '' to
user_home. Another url was added named 'login' in the main urls file;
There may be a potential, however unlikely, problem if a user logs in
twice by entering in the home address twice.