Change Logoout redirect to login page instead of user_home page. Also fixed "user_home/" url.
Changed the logout redirect to from the user_home page to the login
page. Also, fixed a url; In the settings file I just chenged the logout
redirect tag to login instead of home. In the urls file I changed the
url of "user_home" to "user_home/; One minor issue. If a user isn't
logged in and they enter the url "user_home/", they will be redirected
to the user home page and have the ability to see the calendar. This
should be fixed later.