Create request time off form.
Created a new app to handle time off requests and added a button to the
main page to navigate to "request time off" page; The reason I created
a new app is to separate the "time off request" models and forms from
the "user" forms and models. Requests are successfully being stored in
the database; The only issue, which has to be fixed, is connecting the
family objects to the requests in both the models.py and the forms.py.