Connect TimeOffRequest to current family.
The TimeOffRequest wasn't connected to the current volunteer's family,
so we connected it; The family variable was uncommented in
TimeOffRequests.models. Also, the family field was assigned to the
TimeOffRequest model through the view; There are a few possible issues.
1) The User MUST be signed in as a volunteer and that volunteer MUST
have a family in order for the TimeOffRequest to work.
2) If the user isn't signed in and tries to access the TimeOffRequest
page they will get some kind of error and it will not work, or a null
family will be entered into the system, I'm not sure which will happen.