Searching for families and Sending emails.
Resetting forgotten emails now ends an email to you if the email is in the database.
Got to CMPT395/setting.py to setup the EMAIL_HOST_USER and EMAIL_HOST_PASSWORD. You will need to
change some gmail setting such as allowing less secure apps and verifying the site.

The EMAIL_HOST is currently set to gmail but for an actual site you would you something like sendgrid.