Create a list of Time Off Requests in the Django admin page and display each request's status. Also add actions to change the status of each request and remove delete action.
The time off requests needed to be updated to ONLY be able to Approve or
Deny the current requests. There is also an action to change a request
back to pending if the status was changed by mistake; A
TimeOffRequestAdmin class was created in the TimeOffRequest admin.py
file to customize the TimeOffRequests Django admin page. The
TimeOffRequests models were also changed to add a 'status' field; There
shouldn't be any issues with these changes.