Automatically insert the current family into the TimeOffRequest. Also Change the buttons on the sidebar to navigate to the TimeOffRequest page.
The Family objects weren't connecting to the TimeOffRequests so I
connected them. Also I changed the buttons on the sidebar so people
couldn't access the templatetest page; I overrode the '__init__'
function and the 'save' function in the TimeOffRequestForm. In
'__init__' I added a family variable and set a default to it as 'None'
in the constructor. Then in the save function I set the family variable
to the variable passed in to the contrcutor; There are no issues that
I'm aware of.