Add delete button for every TimeOffRequest. Delete the TimeOffRequests from the database. Also change the string representation of Family objects from "user family_name" to just the family name.
There needed to be a delete button for each time off request in case the
user makes a mistake; I added a Form and inside the form a button to
send the TimeOffRequest information to the view, where it is deleted
from the database. I used Declan's process as a guideline; There are no
known issues at this time.