Refactor the field trip setup including creation and deletion. Also change around the urls of the field trip page.
Instead of having a separate page to signup for field trips and viewing
field trips I changed them both into one page and added buttons on the
side to sign up for field trips; I created a table in the html page and
only allowed admins to delete whole field trips. When a field trip is
deleted, all FieldTripSignups associated with that field trip are also
deleted. I added a few lines of code in the fieldTrips.views.py file;
The only issue is the user has to be signed into a volunteer to sign up
for a field trip, otherwise the page will crash.